style: move Deck to new file
